refactor: rename the names of component | SCRUM-41
